Add pusher-js package and enable real-time functionality
This commit adds the 'pusher-js' package to the project and implements real-time functionality by enabling the Laravel Echo and Pusher services in the 'bootstrap.js' file. It also adjusts the 'href' prop in the 'ResponsiveNavLink' Vue component to accept null values as default. The dependencies stated in both package.json and package-lock.json files have been updated accordingly.